+++ lib/lightning_web/controllers/webhooks_controller.ex
@@ -1,8 +1,11 @@
 defmodule LightningWeb.WebhooksController do
   use LightningWeb, :controller
 
-  alias Lightning.Extensions.RateLimiter
-  alias Lightning.Extensions.RateLimiting.Context
+  alias Lightning.Extensions.RateLimiting
+  alias Lightning.Extensions.UsageLimiting.Action
+  alias Lightning.Extensions.UsageLimiting.Context
+  alias Lightning.Services.RateLimiter
+  alias Lightning.Services.UsageLimiter
   alias Lightning.Workflows
   alias Lightning.WorkOrders
 
@@ -12,8 +15,13 @@ defmodule LightningWeb.WebhooksController do
   def create(conn, _params) do
     with %Workflows.Trigger{enabled: true, workflow: %{project_id: project_id}} =
            trigger <- conn.assigns.trigger,
+         {:ok, without_run?} <- check_skip_run_creation(project_id),
          :ok <-
-           RateLimiter.limit_request(conn, %Context{project_id: project_id}, []) do
+           RateLimiter.limit_request(
+             conn,
+             %RateLimiting.Context{project_id: project_id},
+             []
+           ) do
       {:ok, work_order} =
         WorkOrders.create_for(trigger,
           workflow: trigger.workflow,
@@ -22,14 +30,20 @@ defmodule LightningWeb.WebhooksController do
             request: build_request(conn),
             type: :http_request,
             project_id: project_id
-          }
+          },
+          without_run: without_run?
         )
 
       conn |> json(%{work_order_id: work_order.id})
     else
-      {:error, _reason, message} ->
+      {:error, reason, %{text: message}} ->
+        status =
+          if reason == :too_many_requests,
+            do: :too_many_requests,
+            else: :payment_required
+
         conn
-        |> put_status(:too_many_requests)
+        |> put_status(status)
         |> json(%{"error" => message})
 
       nil ->
@@ -43,6 +57,22 @@ defmodule LightningWeb.WebhooksController do
           message:
             "Unable to process request, trigger is disabled. Enable it on OpenFn to allow requests to this endpoint."
         })
+    end
+  end
+
+  defp check_skip_run_creation(project_id) do
+    case UsageLimiter.limit_action(
+           %Action{type: :new_run},
+           %Context{project_id: project_id}
+         ) do
+      :ok ->
+        {:ok, false}
+
+      {:error, :too_many_runs, _message} ->
+        {:ok, true}
+
+      error ->
+        error
     end
   end