+++ lib/lightning_web/live/workflow_live/helpers.ex
@@ -3,7 +3,11 @@ defmodule LightningWeb.WorkflowLive.Helpers do
   Helper functions for the Workflow LiveViews.
   """
 
+  alias Lightning.Extensions.UsageLimiting
+  alias Lightning.Extensions.UsageLimiting.Action
+  alias Lightning.Extensions.UsageLimiting.Context
   alias Lightning.Repo
+  alias Lightning.Services.UsageLimiter
 
   alias Lightning.Workflows
   alias Lightning.WorkOrder
@@ -19,16 +23,28 @@ defmodule LightningWeb.WorkflowLive.Helpers do
           {:ok,
            %{
              workorder: WorkOrder.t(),
-             workflow: Workflows.Workflow.t()
+             workflow: Workflows.Workflow.t(),
+             message: UsageLimiting.message()
            }}
           | {:error, Ecto.Changeset.t(Workflows.Workflow.t())}
           | {:error, Ecto.Changeset.t(WorkOrders.Manual.t())}
+          | {:error, UsageLimiting.message()}
   def save_and_run(workflow_changeset, params, opts) do
     Lightning.Repo.transact(fn ->
-      with {:ok, workflow} <- save_workflow(workflow_changeset),
-           {:ok, manual} <- build_manual_workorder(params, workflow, opts),
-           {:ok, workorder} <- WorkOrders.create_for(manual) do
-        {:ok, %{workorder: workorder, workflow: workflow}}
+      %{id: project_id} = Keyword.fetch!(opts, :project)
+
+      case UsageLimiter.limit_action(%Action{type: :new_run}, %Context{
+             project_id: project_id
+           }) do
+        {:error, _reason, message} ->
+          {:error, message}
+
+        :ok ->
+          with {:ok, workflow} <- save_workflow(workflow_changeset),
+               {:ok, manual} <- build_manual_workorder(params, workflow, opts),
+               {:ok, workorder} <- WorkOrders.create_for(manual) do
+            {:ok, %{workorder: workorder, workflow: workflow}}
+          end
       end
     end)
   end