+++ CHANGELOG.md
@@ -17,6 +17,12 @@ and this project adheres to
 
 ### Added
 
+- Allows limiting creation of new runs and retries.
+  [#1754](https://github.com/OpenFn/Lightning/issues/1754)
+- Add specific messages for log, input, and output tabs when a run is lost
+  [#1757](https://github.com/OpenFn/lightning/issues/1757)
+- Soft and hard limits for runs created by webhook trigger.
+  [#1859](https://github.com/OpenFn/Lightning/issues/1859)
 - Allow admins to set project retention periods
   [#1760](https://github.com/OpenFn/lightning/issues/1760)
 - Automatically wipe input/output data after their retention period
@@ -25,6 +31,8 @@ and this project adheres to
   [#1761](https://github.com/OpenFn/lightning/issues/1761)
 
 ### Changed
+
+- Upgrade Elixir to 1.16.2
 
 ### Fixed
 
@@ -44,6 +52,8 @@ and this project adheres to
 - Automatically create Github workflows in a target repository/branch when users
   set up a Github repo::OpenFn project sync
   [#1046](https://github.com/OpenFn/lightning/issues/1046)
+- Allows limiting creation of new runs and retries.
+  [#1754](https://github.com/OpenFn/Lightning/issues/1754)
 
 ### Changed